feat(builder): base-builder
