+++ bin/builder/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "base-builder"
+name = "base-builder-bin"
 description = "Base Builder Binary"
 
 version.workspace = true
@@ -9,14 +9,18 @@ license.workspace = true
 homepage.workspace = true
 repository.workspace = true
 
+[[bin]]
+name = "base-builder"
+path = "src/main.rs"
+
 [lints]
 workspace = true
 
 [dependencies]
 # Workspace
 base-builder-cli.workspace = true
+base-builder.workspace = true
 base-cli-utils.workspace = true
-op-rbuilder.workspace = true
 
 # CLI
 clap.workspace = true
@@ -26,4 +30,4 @@ reth-cli-util.workspace = true
 
 [features]
 default = [ "jemalloc" ]
-jemalloc = [ "op-rbuilder/jemalloc", "reth-cli-util/jemalloc" ]
+jemalloc = [ "base-builder/jemalloc", "reth-cli-util/jemalloc" ]
