+++ crates/builder/base-builder/src/lib.rs
@@ -7,12 +7,22 @@
 mod execution;
 pub use execution::{ExecutionInfo, TxnExecutionError, TxnOutcome};
 
+mod engine;
+pub use engine::OpEngineApiBuilder;
+
+mod metrics;
+pub use metrics::OpRBuilderMetrics;
+
+mod traits;
+pub use traits::{NodeBounds, PoolBounds, ClientBounds, PayloadTxsBounds, NodeComponents};
+
+mod launcher;
+pub use launcher::{BuilderLauncher, launch};
+
+mod storage;
+pub use storage::{BaseApiExtServer, StoredBackrunBundle, StoreData, TxData, TxDataStore, TxDataStoreExt};
+
 pub mod flashblocks;
-pub mod launcher;
-pub mod metrics;
-pub mod primitives;
-pub mod traits;
-pub mod tx_data_store;
 
 #[cfg(any(test, feature = "test-utils"))]
 pub mod test_utils;
