+++ crates/builder/base-builder/src/storage/store.rs
@@ -1,46 +1,26 @@
+//! Transaction data store.
+//!
+//! Provides a concurrent cache for transaction metadata including resource metering
+//! information and backrun bundles. Uses LRU eviction to bound memory usage.
+
 use std::{
     fmt::Debug,
     sync::{
         Arc,
         atomic::{AtomicBool, Ordering},
     },
-    time::Instant,
 };
 
 use alloy_consensus::Transaction;
-use alloy_primitives::{Address, TxHash};
+use alloy_primitives::TxHash;
 use base_bundles::{AcceptedBundle, MeterBundleResponse};
 use concurrent_queue::ConcurrentQueue;
-use jsonrpsee::{
-    core::{RpcResult, async_trait},
-    proc_macros::rpc,
-};
 use reth_optimism_txpool::OpPooledTransaction;
 use reth_transaction_pool::PoolTransaction;
-use tracing::{debug, info, warn};
-use uuid::Uuid;
+use tracing::{debug, info};
 
-use crate::metrics::OpRBuilderMetrics;
-
-#[derive(Clone, Debug)]
-pub struct StoredBackrunBundle {
-    pub bundle_id: Uuid,
-    pub sender: Address,
-    pub backrun_txs: Vec<OpPooledTransaction>,
-    pub total_priority_fee: u128,
-}
-
-#[derive(Clone, Default, Debug)]
-pub struct TxData {
-    pub metering: Option<MeterBundleResponse>,
-    pub backrun_bundles: Vec<StoredBackrunBundle>,
-}
-
-struct StoreData {
-    by_tx_hash: dashmap::DashMap<TxHash, TxData>,
-    lru: ConcurrentQueue<TxHash>,
-    metering_enabled: AtomicBool,
-}
+use super::{StoredBackrunBundle, StoreData, TxData};
+use crate::OpRBuilderMetrics;
 
 #[derive(Clone)]
 pub struct TxDataStore {
@@ -239,76 +219,8 @@ impl Default for TxDataStore {
     }
 }
 
-#[cfg_attr(not(test), rpc(server, namespace = "base"))]
-#[cfg_attr(test, rpc(server, client, namespace = "base"))]
-pub trait BaseApiExt {
-    #[method(name = "sendBackrunBundle")]
-    async fn send_backrun_bundle(&self, bundle: AcceptedBundle) -> RpcResult<()>;
 
-    #[method(name = "setMeteringInformation")]
-    async fn set_metering_information(
-        &self,
-        tx_hash: TxHash,
-        meter: MeterBundleResponse,
-    ) -> RpcResult<()>;
 
-    #[method(name = "setMeteringEnabled")]
-    async fn set_metering_enabled(&self, enabled: bool) -> RpcResult<()>;
-
-    #[method(name = "clearMeteringInformation")]
-    async fn clear_metering_information(&self) -> RpcResult<()>;
-}
-
-#[derive(Debug)]
-pub struct TxDataStoreExt {
-    store: TxDataStore,
-    metrics: OpRBuilderMetrics,
-}
-
-impl TxDataStoreExt {
-    pub fn new(store: TxDataStore) -> Self {
-        Self { store, metrics: OpRBuilderMetrics::default() }
-    }
-}
-
-#[async_trait]
-impl BaseApiExtServer for TxDataStoreExt {
-    async fn send_backrun_bundle(&self, bundle: AcceptedBundle) -> RpcResult<()> {
-        self.metrics.backrun_bundles_received_total.increment(1);
-
-        let start = Instant::now();
-        self.store.insert_backrun_bundle(bundle).map_err(|e| {
-            warn!(target: "tx_data_store", error = %e, "Failed to store bundle");
-            jsonrpsee::types::ErrorObject::owned(
-                jsonrpsee::types::error::INTERNAL_ERROR_CODE,
-                format!("Failed to store bundle: {e}"),
-                None::<()>,
-            )
-        })?;
-        self.metrics.backrun_bundle_insert_duration.record(start.elapsed().as_secs_f64());
-
-        Ok(())
-    }
-
-    async fn set_metering_information(
-        &self,
-        tx_hash: TxHash,
-        metering: MeterBundleResponse,
-    ) -> RpcResult<()> {
-        self.store.insert_metering(tx_hash, metering);
-        Ok(())
-    }
-
-    async fn set_metering_enabled(&self, enabled: bool) -> RpcResult<()> {
-        self.store.set_metering_enabled(enabled);
-        Ok(())
-    }
-
-    async fn clear_metering_information(&self) -> RpcResult<()> {
-        self.store.clear_metering();
-        Ok(())
-    }
-}
 
 #[cfg(test)]
 mod tests {
@@ -318,6 +230,7 @@ mod tests {
     use alloy_signer_local::PrivateKeySigner;
     use op_alloy_consensus::OpTxEnvelope;
     use op_alloy_rpc_types::OpTransactionRequest;
+    use uuid::Uuid;
 
     use super::*;
