+++ crates/builder/base-builder/src/test_utils/artifacts/genesis.json.tmpl
