+++ crates/builder/base-builder/src/test_utils/artifacts/test-jwt-secret.txt
