+++ crates/builder/base-builder/tests/backrun.rs
@@ -4,7 +4,7 @@ use alloy_eips::eip2718::Encodable2718;
 use alloy_primitives::{TxHash, U256};
 use alloy_provider::Provider;
 use base_bundles::{AcceptedBundle, MeterBundleResponse};
-use op_rbuilder::test_utils::{ChainDriverExt, ONE_ETH, setup_test_instance};
+use base_builder::test_utils::{ChainDriverExt, ONE_ETH, setup_test_instance};
 use uuid::Uuid;
 
 /// Tests that backrun bundles are all-or-nothing:
