+++ crates/builder/base-builder/tests/data_availability.rs
@@ -1,7 +1,7 @@
 #![allow(missing_docs)]
 
 use alloy_provider::Provider;
-use op_rbuilder::test_utils::{
+use base_builder::test_utils::{
     BlockTransactionsExt, ChainDriverExt, TransactionBuilderExt, setup_test_instance,
 };
