+++ crates/builder/base-builder/tests/forks.rs
@@ -4,7 +4,7 @@ use std::time::Duration;
 
 use alloy_eips::{BlockNumberOrTag::Latest, Encodable2718, eip1559::MIN_PROTOCOL_BASE_FEE};
 use alloy_primitives::bytes;
-use op_rbuilder::test_utils::{BlockTransactionsExt, setup_test_instance};
+use base_builder::test_utils::{BlockTransactionsExt, setup_test_instance};
 
 #[tokio::test]
 async fn jovian_block_parameters_set() -> eyre::Result<()> {
