+++ crates/builder/base-builder/tests/ordering.rs
@@ -3,7 +3,7 @@
 use alloy_consensus::Transaction;
 use alloy_network::TransactionResponse;
 use futures::{StreamExt, future::join_all, stream};
-use op_rbuilder::test_utils::{ChainDriverExt, ONE_ETH, setup_test_instance};
+use base_builder::test_utils::{ChainDriverExt, ONE_ETH, setup_test_instance};
 
 /// This test ensures that the transactions are ordered by fee priority within each flashblock.
 /// We expect breaks in global ordering that align with flashblock boundaries.
