--- crates/builder/op-rbuilder/src/primitives/mod.rs
@@ -1,1 +1,0 @@
-pub mod reth;
