--- crates/builder/op-rbuilder/src/primitives/reth/mod.rs
@@ -1,1 +1,0 @@
-pub mod engine_api_builder;
